changed datatype in TextFileWriter and MatFileWriter of cpp runtime
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@21871 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/cpp/Include/Core/HistoryImpl.h
@@ -9,7 +9,7 @@
 using std::map;
 #endif
 
-template< template <unsigned long,unsigned long,unsigned long,unsigned long>  class ResultsPolicy,unsigned long dim_1,unsigned long dim_2,unsigned long dim_3,unsigned long dim_4>
+template< template <size_t,size_t,size_t,size_t>  class ResultsPolicy,size_t dim_1, size_t dim_2,size_t dim_3,size_t dim_4>
 class HistoryImpl: public IHistory,
     public ResultsPolicy<dim_1,dim_2,dim_3,dim_4>
 {
